requirements/dev.txt
-r common.txt
coverage
sphinx
pylint
pycodestyle
django-debug-toolbar
django-extensions
tox